scx_rusty: turn cpumask library into a *.c file and clean it up
